+++ src/Wwa.Api/Ioc/ResolverExtensions.cs
@@ -1,11 +1,6 @@
 // Copyright 2017 (c) [Denis Da Silva]. All rights reserved.
 // See License.txt in the project root for license information.
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Web.Http.Dependencies;
 using Wwa.Core.Ioc;
 
@@ -13,7 +8,7 @@ namespace Wwa.Api.Ioc
 {
     public static class ResolverExtensions
     {
-        public static IDependencyResolver GetHttpResolver(this IResolverManager manager)
+        public static IDependencyResolver GetHttpResolver(this IDependencyManager manager)
         {
             manager.Build();
             return new IocDependencyResolver(manager.Resolver);
