+++ src/Wwa.Core/Ioc/IDependencyManager.cs
@@ -1,0 +1,12 @@
+// Copyright 2017 (c) [Denis Da Silva]. All rights reserved.
+// See License.txt in the project root for license information.
+
+namespace Wwa.Core.Ioc
+{
+    public interface IDependencyManager
+    {
+        IResolver Resolver { get; }
+        void Build();
+        void MapAssembly<AssemblyContainer>();
+    }
+}
