+++ src/Wwa.Ioc.Unity/Wwa.Ioc.Unity.csproj
@@ -52,7 +52,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="UnityResolverManager.cs" />
+    <Compile Include="UnityDependencyManager.cs" />
     <Compile Include="UnityResolver.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
